Add function that will be reused to utilties and implement the new version in traveler route.
