+++ routes/traveler.js
@@ -1,6 +1,7 @@
 /*jslint es5: true*/
 var config = require('../config/config.js');
 var configPath = config.configPath;
+var routesUtilities = require('../utilities/routes.js');
 
 var ad = require('../' + configPath + '/ad.json');
 var ldapClient = require('../lib/ldap-client');
@@ -126,27 +127,6 @@ function cloneTraveler(source, req, res) {
     });
   });
 }
-
-function filterBody(strings) {
-  return function (req, res, next) {
-    var k, found = false;
-    for (k in req.body) {
-      if (req.body.hasOwnProperty(k)) {
-        if (strings.indexOf(k) !== -1) {
-          found = true;
-        } else {
-          req.body[k] = null;
-        }
-      }
-    }
-    if (found) {
-      next();
-    } else {
-      return res.send(400, 'cannot find required information in body');
-    }
-  };
-}
-
 
 function getSharedWith(sharedWith, name) {
   var i;
@@ -624,7 +604,7 @@ module.exports = function (app) {
     });
   });
 
-  app.post('/travelers/', auth.ensureAuthenticated, filterBody(['form', 'source']), function (req, res) {
+  app.post('/travelers/', auth.ensureAuthenticated, routesUtilities.filterBody(['form', 'source']), function (req, res) {
     if (req.body.form) {
       Form.findById(req.body.form, function (err, form) {
         if (err) {
@@ -715,7 +695,7 @@ module.exports = function (app) {
     });
   });
 
-  app.put('/travelers/:id/archived', auth.ensureAuthenticated, filterBody(['archived']), function (req, res) {
+  app.put('/travelers/:id/archived', auth.ensureAuthenticated, routesUtilities.filterBody(['archived']), function (req, res) {
     Traveler.findById(req.params.id, 'createdBy archived').exec(function (err, doc) {
       if (err) {
         console.error(err);
@@ -765,7 +745,7 @@ module.exports = function (app) {
     });
   });
 
-  app.put('/travelers/:id/config', auth.ensureAuthenticated, filterBody(['title', 'description', 'deadline']), function (req, res) {
+  app.put('/travelers/:id/config', auth.ensureAuthenticated, routesUtilities.filterBody(['title', 'description', 'deadline']), function (req, res) {
     Traveler.findById(req.params.id, function (err, doc) {
       var k;
       if (err) {
@@ -860,7 +840,7 @@ module.exports = function (app) {
   });
 
 
-  app.post('/travelers/:id/devices/', auth.ensureAuthenticated, filterBody(['newdevice']), function (req, res) {
+  app.post('/travelers/:id/devices/', auth.ensureAuthenticated, routesUtilities.filterBody(['newdevice']), function (req, res) {
     Traveler.findById(req.params.id, function (err, doc) {
       if (err) {
         console.error(err);
